portal-master/portal-impl/src/com/liferay/portal/fabric/local/worker/LocalFabricWorkerProcessCallableExecutor.java
/**
 * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
 *
 * This library is free software; you can redistribute it and/or modify it under
 * the terms of the GNU Lesser General Public License as published by the Free
 * Software Foundation; either version 2.1 of the License, or (at your option)
 * any later version.
 *
 * This library is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
 * details.
 */

package com.liferay.portal.fabric.local.worker;

import com.liferay.portal.fabric.status.JMXProxyUtil.ProcessCallableExecutor;
import com.liferay.portal.kernel.concurrent.NoticeableFuture;
import com.liferay.portal.kernel.process.ProcessCallable;
import com.liferay.portal.kernel.process.ProcessChannel;

import java.io.Serializable;

/**
 * @author Shuyang Zhou
 */
public class LocalFabricWorkerProcessCallableExecutor
	implements ProcessCallableExecutor {

	public LocalFabricWorkerProcessCallableExecutor(
		ProcessChannel<? extends Serializable> processChannel) {

		_processChannel = processChannel;
	}

	@Override
	public <V extends Serializable> NoticeableFuture<V> execute(
		ProcessCallable<V> processCallable) {

		return _processChannel.write(processCallable);
	}

	private final ProcessChannel<? extends Serializable> _processChannel;

}